babel-plugin: use jest. test for babel v6, v7
