+++ .eslintrc.json
@@ -9,10 +9,15 @@
   "rules": {
   	"comma-dangle": ["error", "always-multiline"]
   },
-  "overrides": {
+  "overrides": [{
     "files": ["**/test/**/*.js"],
     "rules": {
       "require-yield": 0
     }
-  }
+  }, {
+    "files": ["**/babel-plugin-redux-saga/test/*.js"],
+    "env": {
+      "jest": true
+    }
+  }]
 }
